The Thirsty Crow Problem
Consider the thirsty crow story where a thirsty crow identifies a jug with little water. It
puts pebbles into the water to raise the level of water and drinks it. Assume that the
initial reading of the jug is 'm1' ml and the crow can drink water if the level of water has
come to 'm2' ml. There are two categories of pebbles small and big in the field. Small
pebble can raise the level of water by 'x' ml and big pebble can raise the level of water by
'y' ml. There are 'n' small pebbles. Crow prefers to place small pebbles in jug and then
only takes big pebbles. Write an algorithm and the Python code to determine the number
of pebbles required to raise the water to ‘m2’ level. For example, if value of 'm1',
'm2','x','y' and 'n' are 54, 300, 10, 20, 10 respectively then the number of pebbles
required is 13.
Input Format:
Read the initial level of water in jug (in ml)
Read the level of water in jug required for drinking (in ml)
Read the height which small pebble will increase (in ml)
Read the height which big pebble will increase (in ml)
Read the number of small pebbles
Output Format:
Number of big pebbles required