Finding Non Defective Nuts
A bag contains a total of 'T' nuts and bolts. Out of which there are 'N' nuts. In a quality
check, x% of the nuts and y% of bolts in the bag was found to be defective. Design an
algorithm and write a Python code to determine the percentage of non-defective items
in the bag. For example, if T is 200, N is 150, x is 50 and y is 20 then the percentage of
non-defective items in bag is 57.5. Round the answer to two decimal places using the
format function.
Input for the Problem
'T' – total number of nuts and bolts in the bag
'N' – Number of nuts
'x' – Percentage of defective nuts in the bag
'y' – Percentage of defective bolts in the bag
The output of the Problem
Percentage of non-defective items in bag
Boundary Conditions
All inputs > 0
Test Data
Input
200
150
50
20
$END
300
100
40
15
Output
57.50
$END
76.67