Finding Rate of Interest
A finance consultant decides to create an application for the public who wants to decide
on the type of investment they can make. He needs an application, which when given
with the initial amount, no of years of investment ‘n’ and the expected amount after ‘n’
years, should tell the rate of interest based on which the customers will choose their
investments. Assume that the application works with only simple interest calculations.
Can you code for Raman? Use Rate of interest = ((expected amount-initial amount) /
(initial amount * no of years)) *100.
Input Format:
The first input value read is the initial amount the customer has.
The second input value read is the no of years he wants to invest.
The third input value is the expected amount after ‘n’ years.
Output Format:
Print rate of interest